+++ packages/components/src/Autocomplete/Menu.tsx
@@ -1,9 +1,10 @@
-import React, { RefObject, useEffect, useLayoutEffect, useState } from "react";
+import React, { RefObject, useEffect, useState } from "react";
 import classnames from "classnames";
 import { createPortal } from "react-dom";
 import { usePopper } from "react-popper";
 import { useOnKeyDown } from "@jobber/hooks/useOnKeyDown";
-import { useIsMounted } from "@jobber/hooks";
+import { useSafeLayoutEffect } from "@jobber/hooks/useSafeLayoutEffect";
+import { useIsMounted } from "@jobber/hooks/useIsMounted";
 import { AnyOption, Option } from "./Option";
 import styles from "./Autocomplete.css";
 import { Text } from "../Text";
@@ -177,7 +178,7 @@ function useRepositionMenu(attachTo: MenuProps["attachTo"], visible = false) {
     ],
   });
 
-  useLayoutEffect(() => {
+  useSafeLayoutEffect(() => {
     popper?.update?.();
   }, [visible]);