+++ packages/components/src/Combobox/components/ComboboxContent/ComboboxLoadMore/ComboboxLoadMore.tsx
@@ -1,0 +1,23 @@
+import React, { useEffect } from "react";
+import { useInView } from "@jobber/hooks/useInView";
+import styles from "./ComboboxLoadMore.css";
+
+interface ComboboxLoadMoreProps {
+  readonly onLoadMore: () => void;
+}
+
+export function ComboboxLoadMore({ onLoadMore }: ComboboxLoadMoreProps) {
+  const [inViewRef, isInView] = useInView<HTMLDivElement>();
+
+  useEffect(() => {
+    isInView && onLoadMore();
+  }, [isInView]);
+
+  return (
+    <div
+      data-testid="ATL-Combobox-Loadmore-Trigger"
+      ref={inViewRef}
+      className={styles.trigger}
+    />
+  );
+}