+++ packages/hooks/src/index.ts
@@ -6,6 +6,7 @@ export * from "./useFocusTrap";
 export * from "./useFormState";
 export * from "./useInView";
 export * from "./useIsMounted";
+export * from "./useSafeLayoutEffect";
 export * from "./useLiveAnnounce";
 export * from "./useOnKeyDown";
 export * from "./usePasswordStrength";