+++ packages/design/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.54.0](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.53.0...@jobber/design@0.54.0) (2024-02-23)
+
+
+### Features
+
+* Add unscheduled icon ([#1770](https://github.com/GetJobber/atlantis/issues/1770)) ([ca0b83d](https://github.com/GetJobber/atlantis/commit/ca0b83d5489056396a451034a4f540a02ce40089))
+
+
+
+
+
 # [0.53.0](https://github.com/GetJobber/atlantis/compare/@jobber/design@0.52.0...@jobber/design@0.53.0) (2024-01-29)